Revert "Add files via upload"
This reverts commit e579f5e56acb0b4282884cdc170604451dc9d6f6.
