+++ CPU.c
@@ -12,12 +12,14 @@
 //  Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
 //  with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+#include <SDL2/SDL.h>
+
 #include <stdio.h>
 #include "CPU.h"
 #include "m68k.h"
 
 #include "Memory.h"
-
+#include "Chipset.h"
 #include "CIA.h"
 
 
@@ -93,127 +95,6 @@ void cpu_execute(){
     m68k_execute(1);
 
 }
-
-
-
-
-void checkInterrupt(Chipset_t* chipset){
-    
-    int intreqr = chipset->intreqr;
-    
-    
-    if(chipset->intenar & 0x4000){    // if master interrupt switch is enabled
-        
-        if(intreqr == 0){
-            m68k_set_irq(0); // no interrupt
-        }
-        
-        if(intreqr & 0x1){
-            m68k_set_irq(1); // Serial Transmit buffer empty
-        }
-        
-        if(intreqr & 0x2){
-            m68k_set_irq(1); // Disk block finished
-        }
-        
-        if(intreqr & 0x4){
-            m68k_set_irq(1); // Software int
-        }
-        
-        if(intreqr & 0x8){
-            m68k_set_irq(2); // Ports, IO and timers
-        }
-        
-        if(intreqr & 0x10){
-            m68k_set_irq(3); // Copper
-        }
-        
-        if(intreqr & 0x20){
-            m68k_set_irq(3); // VBL
-        }
-        
-        if(intreqr & 0x40){
-            m68k_set_irq(3); // Blitter finished
-        }
-        
-        if(intreqr & 0x80){
-            m68k_set_irq(4); // Audio 0
-        }
-        
-        if(intreqr & 0x100){
-            m68k_set_irq(4); // Audio 1
-        }
-        
-        if(intreqr & 0x200){
-            m68k_set_irq(4); // Audio 2
-        }
-        
-        if(intreqr & 0x400){
-            m68k_set_irq(4); // Audio 3
-        }
-        
-        if(intreqr & 0x800){
-            m68k_set_irq(5); // Serial receive buffer full
-        }
-        
-        if(intreqr & 0x1000){
-            m68k_set_irq(5); // Disk sync
-        }
-        
-        if(intreqr & 0x2000){
-            m68k_set_irq(6); // External Int
-        }
-        
-    }else{
-        m68k_set_irq(0); // no interrupt
-    }
-    
-    return;
-    
-    // The old interrupt handling code... works the same as the above code, no idea which is better :-/
-    
-    if(chipset->intenar & 16384){ // if master interrupt switch is enabled
-        
-        uint16_t intMask = chipset->intreqr & chipset->intenar; //only set the int level, if bits are enabled
-        
-        if(intMask !=0){
-            
-            if(intMask & 8192){  // External int
-                m68k_set_irq(6);
-            }else if(intMask & 4096){ // Disk sync
-                m68k_set_irq(5);
-            }else if(intMask & 2048){ // Serial receive buffer full
-                m68k_set_irq(5);
-            }else if(intMask & 1024){ // Audio 3
-                m68k_set_irq(4);
-            }else if(intMask & 512){  // Audio 2
-                m68k_set_irq(4);
-            }else if(intMask & 256){  // Audio 1
-                m68k_set_irq(4);
-            }else if(intMask & 128){  // Audio 0
-                m68k_set_irq(4);
-            }else if(intMask & 64){   // Blitter finished
-                m68k_set_irq(3);
-            }else if(intMask & 32){   // VBL
-                m68k_set_irq(3);
-            }else if(intMask & 16){   // Copper
-                m68k_set_irq(3);
-            }else if(intMask & 8){    // Ports, IO and timers
-                m68k_set_irq(2);
-            }else if(intMask & 4){    // Software int
-                m68k_set_irq(1);
-            }else if(intMask & 2){    // Disk block finished
-                m68k_set_irq(1);
-            }else if(intMask & 1){    // Serial Transmit buffer empty
-                m68k_set_irq(1);
-            }
-        }
-    }
-    
-}
-
-
-
 
 //Emulator functions
