+++ Memory.c
@@ -14,6 +14,7 @@
 
 #include "Memory.h"
 #include "Kick13.h"
+#include "Kick31.h"
 #include "Chipset.h"
 #include "CIA.h"
